Refactor: update edges further (#38)
* refactor: switch vertices from indices to pointers

Step 1 of the changes I mentioned in #30, just changing `Cell` to use direct pointers to other cells rather than indices.  The other changes will likely affect more files, so I'm starting small.

* refactor: switch edges from indices to pointers

Step 2 of changes mentioned in #30.  Builds on #31.

* refactor: improve styling and implementations

Improved styling by removing whitespace and changing a variable name.
Followed Saad's advice to improve the implementations of `cloneCell` and
the `queen` case of `getPossibleMoves`.

* refactor: switched `getAvailableMoves` to use Set

It used to return `Cell[]`, but now it returns `Set<Cell>`.
I believe I traced the effects of the changes, but as always, it's good
for some people to double-check.

* refactor: updated variable names and misc.

I updated the variable names in utils.ts to be `edge` or `vertex` if
they were a constant, and `currEdge` or `currVertex` if they change over
time.  I also made some miscellaneous changes in response to comments on
the PR.

* Refactor: change edges representation to structs

This change has some downsides, but on the whole it should significantly
increase code readability.

* refactor: cleaned up code related to changes

Also, I switched the knight's code to use a filter. We can talk about
it, but I think it's a bit more readable this way.

* style: format file

* refactor: change variable declaration from let to const

---------

Co-authored-by: Saad Tariq <saadtariq.dev@outlook.com>