docs: improve README instructions
This gives more of the context of how to clone the project and which
directory to be in when running various commands.

I feel like there might be more we can do to make it easier to follow,
so feel free to give suggestions! I could also send it to a friend to
see if they can follow the instructions, do you think that's a good
idea?